23 / 01/ 26 상점 모든 기능 리펙토링 진행

+++ Assets/Scripts/Stat/LivingEntity.cs
@@ -53,7 +53,7 @@ public abstract class LivingEntity : MonoBehaviour
             moveSpeed = value;
         }
     }
-    public float Armor { get; set; }
+    public int Armor { get; set; }
     public int DefaultAttackDamage 
     {
         get { return defaultAttackDamage; }
